Additional LART fixes and more probes
LART was creating a wrong modularity matrix, but now it's using the same one as glouvain. The nodes are mapped from super high numbers down to a smaller range (this gave wrong node id). Added additional probes for instrumentation

+++ multinet/include/community/lart.h
@@ -89,7 +89,8 @@ private:
 
 	void modmat(std::vector<Eigen::SparseMatrix<double>> a, double gamma, Eigen::SparseMatrix<double>& sA);
 
-	vector<long> get_partition(vector<dlib::bu_cluster> clusters, int maxmodix, size_t L, size_t N);
+	vector<unsigned int> get_partition(vector<dlib::bu_cluster> clusters, int maxmodix, size_t L, size_t N);
+
 
 };
